t/tree-test-foreign.dat

#data
<math></math>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <math math>

#data
<math>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <math math>

#data
<math><mi></mi></math>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>

#data
<math><mi>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>

#data
<svg></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|     "x"

#data
<svg>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<svg/>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|     "x"

#data
<svg><textpath></textpath></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg textPath>
|     "x"

#data
<svg><foreignobject/></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|     "x"

#data
<svg><foreignObject></foreignobject></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|     "x"

#data
<svg><foreignObject>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>

#data
<svg><foreignobject></body>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         "x"

#data
<p><svg><foreignobject><address>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <address>

#data
<p><svg><foreignobject><article>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <article>

#data
<p><svg><foreignobject><aside>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <aside>

#data
<p><svg><foreignobject><blockquote>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <blockquote>

#data
<p><svg><foreignobject><center>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <center>

#data
<p><svg><foreignobject><datagrid>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <datagrid>

#data
<p><svg><foreignobject><details>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <details>

#data
<p><svg><foreignobject><dialog>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <dialog>

#data
<p><svg><foreignobject><dir>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <dir>

#data
<p><svg><foreignobject><div>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <div>

#data
<p><svg><foreignobject><dl>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <dl>

#data
<p><svg><foreignobject><figure>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <figure>

#data
<p><svg><foreignobject><footer>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <footer>

#data
<p><svg><foreignobject><h6>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <h6>

#data
<p><svg><foreignobject><header>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <header>

#data
<p><svg><foreignobject><menu>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <menu>

#data
<p><svg><foreignobject><nav>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <nav>

#data
<p><svg><foreignobject><ol>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <ol>

#data
<p><svg><foreignobject><p>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <p>

#data
<p><svg><foreignobject><section>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <section>

#data
<p><svg><foreignobject><ul>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <ul>

#data
<p><svg><foreignobject><pre>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <pre>

#data
<p><svg><foreignobject><listing>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <listing>

#data
<p><svg><foreignobject><form>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <form>

#data
<p><svg><foreignobject><fieldset>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <fieldset>

#data
<p><svg><foreignobject><li>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <li>

#data
<p><svg><foreignobject><dd>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <dd>

#data
<p><svg><foreignobject><dt>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <dt>

#data
<p><svg><foreignobject><plaintext>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <plaintext>

#data
<address><svg><foreignobject></address>x
#errors
no DOCTYPE
unmatched </address>
EOF
#document
| <html>
|   <head>
|   <body>
|     <address>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<article><svg><foreignobject></article>x
#errors
no DOCTYPE
unmatched </article>
EOF
#document
| <html>
|   <head>
|   <body>
|     <article>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<aside><svg><foreignobject></aside>x
#errors
no DOCTYPE
unmatched </aside>
EOF
#document
| <html>
|   <head>
|   <body>
|     <aside>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<ul><svg><foreignobject></ul>x
#errors
no DOCTYPE
unmatched </ul>
EOF
#document
| <html>
|   <head>
|   <body>
|     <ul>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<form><svg><foreignobject></form>x
#errors
no DOCTYPE
unmatched end tag
EOF
#document
| <html>
|   <head>
|   <body>
|     <form>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<p><svg><foreignobject></p>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<dd><svg><foreignobject></dd>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <dd>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<dt><svg><foreignobject></dt>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <dt>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<li><svg><foreignobject></li>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <li>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<h1><svg><foreignobject></h1>x
#errors
no DOCTYPE
unmatched end tag
EOF
#document
| <html>
|   <head>
|   <body>
|     <h1>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<nobr><svg><foreignobject><nobr></nobr></foreignObject></svg></nobr>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <nobr>
|       <svg svg>
|         <svg foreignObject>
|           <nobr>

#data
<button><svg><foreignobject><button></button></foreignObject></svg></button>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <button>
|       <svg svg>
|         <svg foreignObject>
|           <button>

#data
<applet><svg><foreignobject><applet></applet></foreignObject></svg></applet>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <applet>
|       <svg svg>
|         <svg foreignObject>
|           <applet>

#data
<applet><svg><foreignobject></applet>x
#errors
no DOCTYPE
unmatched end tag
EOF
#document
| <html>
|   <head>
|   <body>
|     <applet>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<button><svg><foreignobject></button>x
#errors
no DOCTYPE
unmatched end tag
EOF
#document
| <html>
|   <head>
|   <body>
|     <button>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<marquee><svg><foreignobject></marquee>x
#errors
no DOCTYPE
unmatched end tag
EOF
#document
| <html>
|   <head>
|   <body>
|     <marquee>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<object><svg><foreignobject></object>x
#errors
no DOCTYPE
unmatched end tag
EOF
#document
| <html>
|   <head>
|   <body>
|     <object>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<p><svg><foreignobject><table></table>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <table>

#data
<p><svg><foreignobject><hr>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <hr>

#data
<option><svg><foreignobject></option>x
#errors
no DOCTYPE
unmatched end tag </option>
EOF
#document
| <html>
|   <head>
|   <body>
|     <option>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<optgroup><svg><foreignobject></optgroup>x
#errors
no DOCTYPE
unmatched end tag </optgroup>
EOF
#document
| <html>
|   <head>
|   <body>
|     <optgroup>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<ruby><rt><svg><foreignobject><rp>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       <rt>
|         <svg svg>
|           <svg foreignObject>
|             <rp>

#data
<ruby><rt><svg><foreignobject><rt>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       <rt>
|         <svg svg>
|           <svg foreignObject>
|             <rt>

#data
<table><tr><td><svg><caption>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg caption>
|                 "a"

#data
<table><tr><td><svg><col>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg col>
|                 "a"

#data
<table><tr><td><svg><colgroup>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg colgroup>
|                 "a"

#data
<table><tr><td><svg><tr>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg tr>
|                 "a"

#data
<table><caption><svg><caption>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg caption>
|             "a"

#data
<table><caption><svg><col>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg col>
|             "a"

#data
<table><caption><svg><colgroup>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg colgroup>
|             "a"

#data
<table><caption><svg><thead>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg thead>
|             "a"

#data
<table><caption><svg><tr>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg tr>
|             "a"

#data
<table><caption><svg><td>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg td>
|             "a"

#data
<table><tr><td><svg><g>a</td>b
#errors
no DOCTYPE
unmatched </td>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "ab"

#data
<table><tr><th><svg><g>a</th>b
#errors
no DOCTYPE
unmatched </th>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <svg svg>
|               <svg g>
|                 "ab"

#data
<table><caption><svg><g>a</td>b
#errors
no DOCTYPE
unmatched </td>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><caption><svg><g>a</th>b
#errors
no DOCTYPE
unmatched </th>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><caption><svg><g>a</caption>b
#errors
no DOCTYPE
unmatched </caption>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><tr><td><svg><g>a</caption>b
#errors
no DOCTYPE
unmatched </caption>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "ab"

#data
<table><tr><th><svg><g>a</caption>b
#errors
no DOCTYPE
unmatched </caption>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <svg svg>
|               <svg g>
|                 "ab"

#data
<table><tr><td><svg><g>a</table>b
#errors
no DOCTYPE
unexpected </table>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "ab"

#data
<table><tr><td><svg><g>a</tbody>b
#errors
no DOCTYPE
unmatched </tbody>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "ab"

#data
<table><tr><td><svg><g>a</tr>b
#errors
no DOCTYPE
unmatched </tr>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "ab"

#data
<table><thead><tr><td><svg><g>a</thead>b
#errors
no DOCTYPE
unmatched </thead>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "ab"

#data
<table><tfoot><tr><td><svg><g>a</tfoot>b
#errors
no DOCTYPE
unmatched </tfoot>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "ab"

#data
<table><caption><svg><g>a</table>b
#errors
no DOCTYPE
unexpected </table>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><caption><svg><g>a</tbody>b
#errors
no DOCTYPE
unmatched </tbody>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><caption><svg><g>a</thead>b
#errors
no DOCTYPE
unmatched </thead>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><caption><svg><g>a</tfoot>b
#errors
no DOCTYPE
unmatched </tfoot>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><caption><svg><g>a</tr>b
#errors
no DOCTYPE
unmatched </tr>
EOF
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><tr><svg><g>a</tr>b<!---->
#errors
no DOCTYPE
<svg> in tr
unexpected </tr>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|         <!--  -->
|     <table>
|       <tbody>
|         <tr>

#data
<table><tbody><svg><g>a</tbody>b<!---->
#errors
no DOCTYPE
<svg> in tbody
unexpected </tbody>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|         <!--  -->
|     <table>
|       <tbody>

#data
<table><thead><svg><g>a</thead>b<!---->
#errors
no DOCTYPE
<svg> in thead
unexpected </thead>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|         <!--  -->
|     <table>
|       <thead>

#data
<table><tfoot><svg><g>a</tfoot>b<!---->
#errors
no DOCTYPE
<svg> in tfoot
unexpected </tfoot>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|         <!--  -->
|     <table>
|       <tfoot>

#data
<table><svg><g>a</table>b
#errors
no DOCTYPE
<svg> in table
unexpected </table>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|     <table>

#data
<table><tbody><tr><svg><g>a</tbody>b<!---->
#errors
no DOCTYPE
<svg> in tr
unexpected </table>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|         <!--  -->
|     <table>
|       <tbody>
|         <tr>

#data
<table><thead><tr><svg><g>a</thead>b<!---->
#errors
no DOCTYPE
<svg> in tr
unexpected </thead>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|         <!--  -->
|     <table>
|       <thead>
|         <tr>

#data
<table><tfoot><tr><svg><g>a</tfoot>b<!---->
#errors
no DOCTYPE
<svg> in tr
unexpected </tfoot>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|         <!--  -->
|     <table>
|       <tfoot>
|         <tr>

#data
<table><tbody><tr><svg><g>a</table>b
#errors
no DOCTYPE
<svg> in tr
unexpected </table>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|     <table>
|       <tbody>
|         <tr>

#data
<table><tbody><tr><svg><g>a</tbody>b<!---->
#errors
no DOCTYPE
<svg> in tr
unexpected </tbody>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|         <!--  -->
|     <table>
|       <tbody>
|         <tr>

#data
<table><thead><tr><svg><g>a</thead>b<!---->
#errors
no DOCTYPE
<svg> in tr
unexpected </thead>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|         <!--  -->
|     <table>
|       <thead>
|         <tr>

#data
<table><tfoot><tr><svg><g>a</tfoot>b<!---->
#errors
no DOCTYPE
<svg> in tr
unexpected </tfoot>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "ab"
|         <!--  -->
|     <table>
|       <tfoot>
|         <tr>

#data
<table><colgroup><svg></colgroup><!----></table>
#errors
no DOCTYPE
<svg> in table
unexpected </colgroup>
unexpected </table>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <!--  -->
|     <table>
|       <colgroup>

#data
<!DOCTYPE HTML>
<p xmlns:xlink="http://www.w3.org/1999/xhtml">x</p>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       xmlns:xlink="http://www.w3.org/1999/xhtml"
|       "x"

#data
<!DOCTYPE HTML>
<svg xmlns:xlink="http://www.w3.org/1999/xlink"></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xmlns xlink="http://www.w3.org/1999/xlink"

#data
<!DOCTYPE HTML>
<svg xmlns:xlink="http://www.w3.org/1999/xhtml"></svg>
#errors
bad namespace
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xmlns xlink="http://www.w3.org/1999/xhtml"

#data
<!DOCTYPE HTML>
<svg xmlns:xlink="http://www.w3.org/1999/xlink "></svg>
#errors
bad namespace
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xmlns xlink="http://www.w3.org/1999/xlink "

#data
<!DOCTYPE html>
<svg><font></font></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg font>

#data
<!DOCTYPE html>
<svg><font>abc</font></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg font>
|         "abc"

#data
<!DOCTYPE html>
<svg><font color=red></font>
#errors
no </svg>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <font>
|       color="red"

#data
<!DOCTYPE HTML><svg><font>x</font>y</svg>z
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg font>
|         "x"
|       "y"
|     "z"

#data
<!DOCTYPE HTML><svg><a>x</a>y</svg>z
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg a>
|         "x"
|       "y"
|     "z"

#data
<!DOCTYPE HTML><svg><c>x</c>y</svg>z
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg c>
|         "x"
|       "y"
|     "z"

#data
<!DOCTYPE html>
<svg><font size=1></font>
#errors
no </svg>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <font>
|       size="1"

#data
<!DOCTYPE html>
<svg><font face="Arial"></font>
#errors
no </svg>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <font>
|       face="Arial"

#data
<!DOCTYPE html>
<svg><font point-size=14></font></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg font>
|         point-size="14"

#data
<!DOCTYPE html>
<table>
  <tr>
    <td>
      <svg>
        <foreignObject>
          <table>
            <td>
          </table>
        </foreignObject>
      </svg>
    <td>
      Hi
  </tr>
</table>
there
#errors
9:0:missing start tag:tr;
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       "
|   "
|       <tbody>
|         <tr>
|           "
|     "
|           <td>
|             "
|       "
|             <svg svg>
|               "
|         "
|               <svg foreignObject>
|                 "
|           "
|                 <table>
|                   "
|             "
|                   <tbody>
|                     <tr>
|                       <td>
|                         "
|           "
|                 "
|         "
|               "
|       "
|             "
|     "
|           <td>
|             "
|       Hi
|   "
|         "
| "
|     "
| there"

#data
<!DOCTYPE HTML><table><tr><td><svg><td>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>

#data
<!DOCTYPE HTML><table><tr><td><svg><td>a
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|                 "a"

#data
<!DOCTYPE HTML><table><tr><td><svg><td></td>b
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|               "b"

#data
<!DOCTYPE HTML><table><tr><td><svg><td></td><td>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|               <svg td>

#data
<!DOCTYPE HTML><table><tr><td><svg><td>a</td>b<td>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|                 "a"
|               "b"
|               <svg td>

#data
<!DOCTYPE HTML><table><tr><td><svg><td><tr>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|                 <svg tr>

#data
<!DOCTYPE HTML><table><tr><td><svg><td></td><tr>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|               <svg tr>

#data
<!DOCTYPE HTML><table><tr><td><svg><td></td><tr></td> <td>
#errors
unexpected </td>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|               <svg tr>
|                 " "
|                 <svg td>

#data
<!DOCTYPE HTML><table><tr><td><svg><td><tr></td> <td>
#errors
no </tr>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|                 <svg tr>
|               " "
|               <svg td>

#data
<!DOCTYPE HTML><table><tr><td><svg><td></tbody><tr>
#errors
unexpected </tbody>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|                 <svg tr>

#data
<!DOCTYPE HTML><table><tr><td><svg><td><tbody>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|                 <svg tbody>

#data
<!DOCTYPE HTML><table><tr><td><svg><td></tr><tbody>
#errors
unexpected </tr>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|                 <svg tbody>

#data
<!DOCTYPE HTML><table><tr><td><svg><td></td><tbody>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|               <svg tbody>

#data
<!DOCTYPE HTML><table><tr><td><svg><td></table>a
#errors
unexpected </table>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg td>
|                 "a"

#data
<!DOCTYPE html><table><tr><td><svg><tbody></tbody>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg tbody>

#data
<!DOCTYPE html><table><tr><td><svg><tbody></tbody>a
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg tbody>
|               "a"

#data
<!DOCTYPE HTML><svg><desc><math><b>
#errors
no </math>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg desc>
|         <math math>
|         <b>

#data
<!DOCTYPE HTML><svg><title><math><b>
#errors
no </math>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg title>
|         <math math>
|         <b>

#data
<!DOCTYPE html><math><mi><svg><b>
#errors
no </svg>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <svg svg>
|         <b>

#data
<!DOCTYPE html><math><mo><svg><b>
#errors
no </svg>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mo>
|         <svg svg>
|         <b>

#data
<!DOCTYPE html><math><mn><svg><b>
#errors
no </svg>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mn>
|         <svg svg>
|         <b>

#data
<!DOCTYPE html><math><ms><svg><b>
#errors
no </svg>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math ms>
|         <svg svg>
|         <b>

#data
<!DOCTYPE html><math><mtext><svg><b>
#errors
no </svg>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mtext>
|         <svg svg>
|         <b>

#data
<!DOCTYPE html><svg><foreignobject><math><b>
#errors
no </math>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <math math>
|         <b>

#data
<!DOCTYPE html><svg><foreignobject><span><math><b>
#errors
no </math> before <b>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <span>
|           <math math>
|           <b>

#data
<!DOCTYPE html><svg><foreignobject><span><math><sub>
#errors
no </math> before <sub>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <span>
|           <math math>
|           <sub>

#data
<!DOCTYPE html><svg><title><mglyph>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg title>
|         <mglyph>

#data
<!DOCTYPE html><svg><title><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg title>
|         <malignmark>

#data
<!DOCTYPE html><svg><desc><mglyph>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg desc>
|         <mglyph>

#data
<!DOCTYPE html><svg><desc><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg desc>
|         <malignmark>

#data
<!DOCTYPE html><svg><foreignObject><mglyph>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <mglyph>

#data
<!DOCTYPE html><svg><foreignObject><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <malignmark>

#data
<!DOCTYPE html><math><annotation-xml><mglyph>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         <math mglyph>

#data
<!DOCTYPE html><math><annotation-xml encoding="text/html"><mglyph>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html"
|         <mglyph>

#data
<!DOCTYPE html><math><annotation-xml><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         <math malignmark>

#data
<!DOCTYPE html><math><annotation-xml encoding="text/html"><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html"
|         <malignmark>

#data
<!DOCTYPE html><math><mi><mglyph>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <math mglyph>

#data
<!DOCTYPE html><math><mi><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <math malignmark>

#data
<!DOCTYPE html><math><mo><mglyph>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mo>
|         <math mglyph>

#data
<!DOCTYPE html><math><mo><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mo>
|         <math malignmark>

#data
<!DOCTYPE html><math><ms><mglyph>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math ms>
|         <math mglyph>

#data
<!DOCTYPE html><math><ms><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math ms>
|         <math malignmark>

#data
<!DOCTYPE html><math><mn><mglyph>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mn>
|         <math mglyph>

#data
<!DOCTYPE html><math><mn><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mn>
|         <math malignmark>

#data
<!DOCTYPE html><math><mtext><mglyph>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mtext>
|         <math mglyph>

#data
<!DOCTYPE html><math><mtext><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mtext>
|         <math malignmark>

#data
<!DOCTYPE html><ruby><math><mi><rt>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <math math>
|         <math mi>
|           <rt>

#data
<!DOCTYPE html><ruby><math><mo><rt>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <math math>
|         <math mo>
|           <rt>

#data
<!DOCTYPE html><ruby><math><mn><rt>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <math math>
|         <math mn>
|           <rt>

#data
<!DOCTYPE html><ruby><math><ms><rt>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <math math>
|         <math ms>
|           <rt>

#data
<!DOCTYPE html><ruby><math><mtext><rt>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <math math>
|         <math mtext>
|           <rt>

#data
<!DOCTYPE html><ruby><svg><desc><rt>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <svg svg>
|         <svg desc>
|           <rt>

#data
<!DOCTYPE html><ruby><svg><title><rt>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <svg svg>
|         <svg title>
|           <rt>

#data
<!DOCTYPE html><ruby><math><annotation-xml><rt>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <math math>
|         <math annotation-xml>
|           <math rt>

#data
<!DOCTYPE html><ruby><math><annotation-xml encoding="text/html"><rt>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <math math>
|         <math annotation-xml>
|           encoding="text/html"
|           <rt>

#data
<!DOCTYPE html><ul><li><math><mi><li>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <math math>
|           <math mi>
|             <li>

#data
<!DOCTYPE html><ul><li><math><mo><li>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <math math>
|           <math mo>
|             <li>

#data
<!DOCTYPE html><ul><li><math><mn><li>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <math math>
|           <math mn>
|             <li>

#data
<!DOCTYPE html><ul><li><math><ms><li>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <math math>
|           <math ms>
|             <li>

#data
<!DOCTYPE html><ul><li><math><mtext><li>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <math math>
|           <math mtext>
|             <li>

#data
<!DOCTYPE html><ul><li><svg><desc><li>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <svg svg>
|           <svg desc>
|             <li>

#data
<!DOCTYPE html><ul><li><svg><title><li>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <svg svg>
|           <svg title>
|             <li>

#data
<!DOCTYPE html><ul><li><math><annotation-xml><li>
#errors
no </annotation-xml></math> before <li>
no </ul> before EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <math math>
|           <math annotation-xml>
|       <li>

#data
<!DOCTYPE html><math><annotation-xml><wbr><rt><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         <math wbr>
|           <math rt>
|             <math malignmark>

#data
<!DOCTYPE html><math><annotation-xml encoding="text/xml"><wbr><rt><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/xml"
|         <math wbr>
|           <math rt>
|             <math malignmark>

#data
<!DOCTYPE html><math><annotation-xml encoding="text/html "><wbr><rt><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html "
|         <math wbr>
|           <math rt>
|             <math malignmark>

#data
<!DOCTYPE html><math><annotation-xml encoding="text/html"><wbr><rt><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html"
|         <wbr>
|         <rt>
|           <malignmark>

#data
<!DOCTYPE html><math><annotation-xml encoding="TEXT/HTML"><wbr><rt><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="TEXT/HTML"
|         <wbr>
|         <rt>
|           <malignmark>

#data
<!DOCTYPE html><math><annotation-xml encoding="application/xhtml+xml"><wbr><rt><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="application/xhtml+xml"
|         <wbr>
|         <rt>
|           <malignmark>

#data
<!DOCTYPE html><math><annotation-xml encoding="application/xhtml+XML"><wbr><rt><malignmark>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="application/xhtml+XML"
|         <wbr>
|         <rt>
|           <malignmark>

#data
<!DOCTYPE html><math><mi><object><span></span></object></mi><mi>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <object>
|           <span>
|       <math mi>

#data
<!DOCTYPE html><p><math><foo><mi><p>a
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <math math>
|         <math foo>
|           <math mi>
|             <p>
|               "a"

#data
<!DOCTYPE html><p><math><mi><xmp>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <math math>
|         <math mi>
|           <xmp>

#data
<!DOCTYPE html><p><math><mi><dd>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <math math>
|         <math mi>
|           <dd>

#data
<!DOCTYPE html><button><math><mi><button>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <button>
|       <math math>
|         <math mi>
|           <button>

#data
<!DOCTYPE html><svg><body x><desc></html><!---->
#errors
no </svg> before <body>
second <body>
no </desc> before </html>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     x=""
|     <svg svg>
|     <desc>
| <!--  -->

#data
<!DOCTYPE html><table><tr><td><svg><desc><td><!---->
#errors
<td> outside of <table>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg desc>
|                 <!--  -->

#data
<!DOCTYPE html><svg><p><desc><form>
#errors
no </svg> before <p>
no </desc> before <form>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <p>
|       <desc>
|     <form>

#data
<!DOCTYPE html><p><svg><desc><table><td>
#errors
no <tr> before <td>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg desc>
|           <table>
|             <tbody>
|               <tr>
|                 <td>

#data
<!DOCTYPE html><svg><script></script></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg script>

#data
<!DOCTYPE html><svg><script>a</script></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg script>
|         "a"

#data
<!DOCTYPE html><svg><g><script></script></g></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         <svg script>

#data
<!DOCTYPE html><math><g><script></script></g></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math g>
|         <math script>

#data
<!DOCTYPE html><svg><script><g></g></script></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg script>
|         <svg g>

#data
<!DOCTYPE html><svg><title><script><g></g></script></title></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg title>
|         <script>
|           "<g></g>"

#data
<!DOCTYPE html><svg><script><g></script></svg>
#errors
no </g>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg script>
|         <svg g>

#data
<!DOCTYPE html><math><script><g></script></math>
#errors
no </g>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math script>
|         <math g>

#data
<!DOCTYPE html><math><annotation-xml><b>
#errors
no </annotation-xml></math> before <b>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|     <b>

#data
<!DOCTYPE html><math><annotation-xml encoding="text/xml"><b>
#errors
no </annotation-xml></math> before <b>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/xml"
|     <b>

#data
<!DOCTYPE html><math><annotation-xml encoding="text/html "><b>
#errors
no </annotation-xml></math> before <b>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html "
|     <b>

#data
<!DOCTYPE html><math><annotation-xml encoding="text/html"><b>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html"
|         <b>

#data
<!DOCTYPE html><math><annotation-xml encoding="application/xhtml+xml"><b>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="application/xhtml+xml"
|         <b>

#data
<!DOCTYPE html><math><annotation-xml encoding="text/html"><x><b>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html"
|         <x>
|           <b>

#data
<!DOCTYPE html><math><annotation-xml><math><b>
#errors
no </math></annotation-xml></math>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         <math math>
|     <b>

#data
<!DOCTYPE html><math><annotation-xml encoding="MathML"><math><b>
#errors
no </math></annotation-xml></math>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="MathML"
|         <math math>
|     <b>

#data
<!DOCTYPE html><math><annotation-xml encoding="text/html"><math><b>
#errors
no </math>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html"
|         <math math>
|         <b>

#data
<!DOCTYPE html><math><annotation-xml encoding="Text/HTML"><math><b>
#errors
no </math>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="Text/HTML"
|         <math math>
|         <b>

#data
<!DOCTYPE html><math><annotation-xml encoding="application/xhtml+xml"><math><b>
#errors
no </math>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="application/xhtml+xml"
|         <math math>
|         <b>

#data
<!DOCTYPE html><math><annotation-xml encoding="Application/xhtml+xml"><math><b>
#errors
no </math>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="Application/xhtml+xml"
|         <math math>
|         <b>

#data
<!DOCTYPE html><math><mi><i>a</math><math>b
#errors
unexpected </math>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|           <math math>
|             "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><math>b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math math>
|         "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><x>b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math x>
|         "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><mi>b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math mi>
|         "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><ms>b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math ms>
|         "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><mo>b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math mo>
|         "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><mn>b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math mn>
|         "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><mtext>b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math mtext>
|         "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><mi><div>b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math mi>
|         <div>
|           <i>
|             "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><mi><b>b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math mi>
|         <i>
|           <b>
|             "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><annotation-xml>b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math annotation-xml>
|         "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><annotation-xml encoding="text/html">b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math annotation-xml>
|         encoding="text/html"
|         <i>
|           "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><annotation-xml encoding="application/xhtml+XML">b
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math annotation-xml>
|         encoding="application/xhtml+XML"
|         <i>
|           "b"

#data
<!DOCTYPE html><math><mi><i>a</mi><annotation-xml encoding="text/html"><svg><desc>x
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math annotation-xml>
|         encoding="text/html"
|         <i>
|           <svg svg>
|             <svg desc>
|               "x"

#data
<!DOCTYPE html><math><mi><i>a</mi><annotation-xml encoding="text/html"><svg><title>x
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math annotation-xml>
|         encoding="text/html"
|         <i>
|           <svg svg>
|             <svg title>
|               "x"

#data
<!DOCTYPE html><math><mi><i>a</mi><annotation-xml encoding="text/html"><svg><foreignobject>x
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <i>
|           "a"
|       <math annotation-xml>
|         encoding="text/html"
|         <i>
|           <svg svg>
|             <svg foreignObject>
|               "x"

#data
<!DOCTYPE html><svg><title><i>a</title><desc>x
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg title>
|         <i>
|           "a"
|       <svg desc>
|         <i>
|           "x"

#data
<!DOCTYPE html><svg><title><i>a</title><title>x
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg title>
|         <i>
|           "a"
|       <svg title>
|         <i>
|           "x"

#data
<!DOCTYPE html><svg><title><i>a</title><foreignObJect>x
#errors
no </i>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg title>
|         <i>
|           "a"
|       <svg foreignObject>
|         <i>
|           "x"

#data
<!DOCTYPE html><math><mi><frameset>
#errors
<frameset> in body
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE html><math><mi>\uFFFD<frameset>
#errors
<frameset> in body
EOF
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "\uFFFD"

#data escaped
<!DOCTYPE html><math><mi><div>\uFFFD<frameset>
#errors
<frameset> in body
EOF
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE html><math><annotation-xml>\uFFFD<mi><frameset>
#errors
<frameset> in body
EOF
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         "\uFFFD"
|         <math mi>

#data escaped
<!DOCTYPE html><math><annotation-xml encoding="text/html"><div>\uFFFD<frameset>
#errors
<frameset> in body
EOF
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE html><math><annotation-xml encoding="application/xhtml+xml"><div>\uFFFD<frameset>
#errors
<frameset> in body
EOF
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE html><svg><desc>\uFFFD<frameset>
#errors
<frameset> in body
EOF
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE html><svg><desc><div>\uFFFD<frameset>
#errors
<frameset> in body
EOF
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE html><svg><title>\uFFFD<frameset>
#errors
<frameset> in body
EOF
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE html><svg><foreignobject>\uFFFD<frameset>
#errors
<frameset> in body
EOF
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE html><svg></svg><![CDATA[foo]]>
#errors
bad comment
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <!-- [CDATA[foo]] -->

#data
<!DOCTYPE html><svg><foreignObject><![CDATA[foo]]></foreignObject></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         "foo"

#data
<!DOCTYPE html><svg><foreignObject><div><![CDATA[foo]]></div></foreignObject></svg>
#errors
bad comment
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <div>
|           <!-- [CDATA[foo]] -->

#data
<!DOCTYPE html><table><svg><desc><i>a</desc><title><b>b</title><foreignobject><table><td>c</table>d</i>e</foreignObject></svg><math><mi>f<i>g</mi><annotation-xml encoding=text/html>h</annotation-xml><annotation-xml>i</annotation-xml></math>j
#errors
<svg> in table
no </i> before </desc>
no </b> before </title>
no <tr> before <td>
no </b> before </i>; AAA
no </b> before </foreignObject>
<math> in table
no </i> before </annotation-xml>
no </i> before </mi>
"j" in table
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg desc>
|         <i>
|           "a"
|       <svg title>
|         <i>
|           <b>
|             "b"
|       <svg foreignObject>
|         <table>
|           <tbody>
|             <tr>
|               <td>
|                 "c"
|         <i>
|           <b>
|             "d"
|         <b>
|           "e"
|     <b>
|       <math math>
|         <math mi>
|           "f"
|           <i>
|             "g"
|         <math annotation-xml>
|           encoding="text/html"
|           <i>
|             "h"
|         <math annotation-xml>
|           "i"
|       <i>
|         "j"
|     <table>
